Use base class and create a generic way of running against command
